requirements.txt
browserforge==1.2.3
camoufox==0.4.11
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Cython==3.0.12
greenlet==3.1.1
idna==3.10
language-tags==1.2.0
lxml==5.3.1
numpy==2.2.4
orjson==3.10.15
platformdirs==4.3.7
playwright==1.51.0
pyee==12.1.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
screeninfo==0.8.1
tqdm==4.67.1
typing_extensions==4.12.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
